[FEAT]::Criando os Event DTOs da aplicação
