+++ src/main/java/com/rafaelgalvao/eventmanager/dto/Event/EventResponseDTO.java
@@ -1,23 +1,11 @@
 package com.rafaelgalvao.eventmanager.dto.Event;
 
-import jakarta.validation.constraints.FutureOrPresent;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-
 import java.time.LocalDateTime;
 
-public record UpdateEventDTO(
-
-        @NotBlank
-        @Size(max = 100)
+public record EventResponseDTO(
+        Long id,
         String title,
-
-        @Size(max = 1000)
         String description,
-
-        @FutureOrPresent
         LocalDateTime eventDate,
-
-        @Size(max = 200)
         String location
 ) {}
